+++ plugins/foreign-db-extractor/package.json
@@ -42,8 +42,10 @@
   "license": "ISC",
   "dependencies": {
     "@flatfile/api": "^1.7.1",
-    "@flatfile/listener": "^1.0.0",
     "mssql": "^10.0.1",
     "node-fetch": "^3.3.2"
+  },
+  "peerDependencies": {
+    "@flatfile/listener": "^1.0.1"
   }
 }
